+++ python/boomer/boosting/cpp/example_wise_rule_evaluation.h
@@ -35,9 +35,10 @@ namespace boosting {
              * gradients and Hessians that are stored in the arrays `totalSumsOfGradients` and `sumsOfGradients` and
              * `totalSumsOfHessians` and `sumsOfHessians`, respectively.
              *
-             * @param labelIndices          A pointer to an array of type `intp`, shape `(prediction.numPredictions_)`,
-             *                              representing the indices of the labels for which the rule should predict or
-             *                              NULL, if the rule should predict for all labels
+             * @param labelIndices          A pointer to an array of type `uint32`, shape
+             *                              `(prediction.numPredictions_)`, representing the indices of the labels for
+             *                              which the rule should predict or NULL, if the rule should predict for all
+             *                              labels
              * @param totalSumsOfGradients  A pointer to an array of type `float64`, shape `(num_labels), representing
              *                              the total sums of gradients for individual labels
              * @param sumsOfGradients       A pointer to an array of type `float64`, shape
@@ -57,7 +58,7 @@ namespace boosting {
              * @param prediction            A pointer to an object of type `LabelWisePredictionCandidate` that should be
              *                              used to store the predicted scores and quality scores
              */
-            virtual void calculateLabelWisePrediction(const intp* labelIndices, const float64* totalSumsOfGradients,
+            virtual void calculateLabelWisePrediction(const uint32* labelIndices, const float64* totalSumsOfGradients,
                                                       float64* sumsOfGradients, const float64* totalSumsOfHessians,
                                                       float64* sumsOfHessians, bool uncovered,
                                                       LabelWisePredictionCandidate* prediction);
@@ -71,9 +72,10 @@ namespace boosting {
              * gradients and Hessians that are stored in the arrays `totalSumsOfGradients` and `sumsOfGradients` and
              * `totalSumsOfHessians` and `sumsOfHessians`, respectively.
              *
-             * @param labelIndices          A pointer to an array of type `intp`, shape `(prediction.numPredictions_)`,
-             *                              representing the indices of the labels for which the rule should predict or
-             *                              NULL, if the rule should predict for all labels
+             * @param labelIndices          A pointer to an array of type `uint32`, shape
+             *                              `(prediction.numPredictions_)`, representing the indices of the labels for
+             *                              which the rule should predict or NULL, if the rule should predict for all
+             *                              labels
              * @param totalSumsOfGradients  A pointer to an array of type `float64`, shape `(num_labels), representing
              *                              the total sums of gradients for individual labels
              * @param sumsOfGradients       A pointer to an array of type `float64`, shape
@@ -112,7 +114,7 @@ namespace boosting {
              * @param prediction            A pointer to an object of type `PredictionCandidate` that should be used to
              *                              store the predicted scores and quality score
              */
-            virtual void calculateExampleWisePrediction(const intp* labelIndices, const float64* totalSumsOfGradients,
+            virtual void calculateExampleWisePrediction(const uint32* labelIndices, const float64* totalSumsOfGradients,
                                                         float64* sumsOfGradients, const float64* totalSumsOfHessians,
                                                         float64* sumsOfHessians, float64* tmpGradients,
                                                         float64* tmpHessians, int dsysvLwork, float64* dsysvTmpArray1,
@@ -152,12 +154,12 @@ namespace boosting {
 
             ~RegularizedExampleWiseRuleEvaluationImpl();
 
-            void calculateLabelWisePrediction(const intp* labelIndices, const float64* totalSumsOfGradients,
+            void calculateLabelWisePrediction(const uint32* labelIndices, const float64* totalSumsOfGradients,
                                               float64* sumsOfGradients, const float64* totalSumsOfHessians,
                                               float64* sumsOfHessians, bool uncovered,
                                               LabelWisePredictionCandidate* prediction) override;
 
-            void calculateExampleWisePrediction(const intp* labelIndices, const float64* totalSumsOfGradients,
+            void calculateExampleWisePrediction(const uint32* labelIndices, const float64* totalSumsOfGradients,
                                                 float64* sumsOfGradients, const float64* totalSumsOfHessians,
                                                 float64* sumsOfHessians, float64* tmpGradients, float64* tmpHessians,
                                                 int dsysvLwork, float64* dsysvTmpArray1, int* dsysvTmpArray2,
