+++ python/boomer/boosting/cpp/example_wise_statistics.h
@@ -31,7 +31,7 @@ namespace boosting {
 
             intp numPredictions_;
 
-            const intp* labelIndices_;
+            const uint32* labelIndices_;
 
             intp numLabels_;
 
@@ -76,9 +76,9 @@ namespace boosting {
              * @param lapackPtr             A shared pointer to an object of type `Lapack` that allows to execute
              *                              different lapack routines
              * @param numPredictions        The number of labels to be considered by the search
-             * @param labelIndices          A pointer to an array of type `intp`, shape `(numPredictions)`, representing
-             *                              the indices of the labels that should be considered by the search or NULL,
-             *                              if all labels should be considered
+             * @param labelIndices          A pointer to an array of type `uint32`, shape `(numPredictions)`,
+             *                              representing the indices of the labels that should be considered by the
+             *                              search or NULL, if all labels should be considered
              * @param numLabels             The total number of labels
              * @param gradients             A pointer to an array of type `float64`, shape `(num_examples, num_labels)`,
              *                              representing the gradients for each example
@@ -95,7 +95,7 @@ namespace boosting {
              */
             DenseExampleWiseRefinementSearchImpl(std::shared_ptr<AbstractExampleWiseRuleEvaluation> ruleEvaluationPtr,
                                                  std::shared_ptr<Lapack> lapackPtr, intp numPredictions,
-                                                 const intp* labelIndices, intp numLabels, const float64* gradients,
+                                                 const uint32* labelIndices, intp numLabels, const float64* gradients,
                                                  const float64* totalSumsOfGradients, const float64* hessians,
                                                  const float64* totalSumsOfHessians);
 
@@ -196,7 +196,7 @@ namespace boosting {
 
             void updateCoveredStatistic(intp statisticIndex, uint32 weight, bool remove) override;
 
-            AbstractRefinementSearch* beginSearch(intp numLabelIndices, const intp* labelIndices) override;
+            AbstractRefinementSearch* beginSearch(intp numLabelIndices, const uint32* labelIndices) override;
 
             void applyPrediction(intp statisticIndex, Prediction* prediction) override;
