+++ python/boomer/boosting/label_wise_statistics.pxd
@@ -18,7 +18,7 @@ cdef extern from "cpp/label_wise_statistics.h" namespace "boosting" nogil:
         # Constructors:
 
         DenseLabelWiseRefinementSearchImpl(shared_ptr[AbstractLabelWiseRuleEvaluation] ruleEvaluationPtr,
-                                           intp numPredictions, const intp* labelIndices, intp numLabels,
+                                           intp numPredictions, const uint32* labelIndices, intp numLabels,
                                            const float64* gradients, const float64* totalSumsOfGradients,
                                            const float64* hessians, const float64* totalSumsOfHessians) except +
